Move router functionality to own component app-routing.module; Create resolver for app-routing;
