+++ src/app/app.module.ts
@@ -1,8 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
-import { RouterModule, Routes } from '@angular/router';
-
+import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HeaderComponent } from './header/header.component';
 import { MainComponent } from './main/main.component';
@@ -20,15 +19,6 @@ import { ErrorComponent } from './error/error.component';
 import {HttpService} from './http.service';
 import { FilterComponent } from './filter/filter.component';
 import { PaginationComponent } from './pagination/pagination.component';
-
-const appRoutes: Routes = [
-  { path: '', component: HomeComponent, pathMatch: 'full' },
-  { path: 'docs', component: DocsComponent },
-  { path: 'examples', component: ExamplesComponent },
-  { path: 'about', component: AboutComponent },
-  { path: 'news', component: NewsComponent },
-  { path: '**', component: PageNotFoundComponent }
-];
 
 @NgModule({
   declarations: [
@@ -50,12 +40,9 @@ const appRoutes: Routes = [
     PaginationComponent
   ],
   imports: [
+    AppRoutingModule,
     BrowserModule,
-    HttpClientModule,
-    RouterModule.forRoot(
-      appRoutes,
-      { enableTracing: true }
-    )
+    HttpClientModule
   ],
   providers: [HttpService],
   bootstrap: [AppComponent]
